refactor: finish Slice B lib splits for three near-cap modules (#4310)
Extract phase-plan outcome defaults/toString, diplomacy declare-war
shortcuts, and treasury emit-input lock-recovery shaping for durable
headroom under the 400-line source ceiling. No AI semantics change.

Refs #4310
